requirements.txt
cryptography
requests
discord
python-decouple